velocidad luz en un medio

+++ Funciones/energía_interna_gases.py
@@ -1,8 +1,8 @@
 
 def energía_interna_gases():
     from Constantes.generales import gene
-    n = int(input("\n¿Número de moles?\n"))
-    T = int(input("\n¿Temperatura?\n"))
+    n = float(input("\n¿Número de moles?\n"))
+    T = float(input("\n¿Temperatura?\n"))
     U = 1.5*n*T*gene["gases ideales"]
     print(U)
