+++ Funciones/transformación_lorentz_posición.py
@@ -2,9 +2,9 @@
 def transformación_lorentz_posición():
     import math
     from Constantes.generales import gene
-    x = int(input("\n¿Posición según el que está en reposo?\n"))
-    v = int(input("\n¿Velocidad del sistema en movimiento?\n"))
-    t = int(input("\n¿Tiempo medido desde el sistema quieto?\n"))
+    x = float(input("\n¿Posición según el que está en reposo?\n"))
+    v = float(input("\n¿Velocidad del sistema en movimiento?\n"))
+    t = float(input("\n¿Tiempo medido desde el sistema quieto?\n"))
     if v < gene["velocidad luz"]:
         xx = (x - v*t)/(math.sqrt(1-(v/gene["velocidad luz"])**2))
         print(xx)
